Handle network errors on symptom report submission (#208)
If an exception is thrown while attempting to submit a symptom report,
the symptom report widget will be rebuilt with questions pre-populated
with the answers from the previous submission attempt. The last step of
the symptom report will be shown after a network error, and an error
snackbar will be displayed notifying the user that something went wrong.